update: 2000并发passed, 不要用windows版本redis测试

+++ mqtt-broker/src/main/java/cn/wizzer/iot/mqtt/server/broker/protocol/DisConnect.java
@@ -41,7 +41,7 @@ public class DisConnect {
     public void processDisConnect(ChannelContext channel, MqttMessage msg) {
         String clientId = (String) channel.getAttribute("clientId");
         SessionStore sessionStore = sessionStoreService.get(clientId);
-        if (sessionStore.isCleanSession()) {
+        if (sessionStore != null && sessionStore.isCleanSession()) {
             subscribeStoreService.removeForClient(clientId);
             dupPublishMessageStoreService.removeByClient(clientId);
             dupPubRelMessageStoreService.removeByClient(clientId);
